extract & use instrumentation helpers
