+++ packages/zipkin/src/instrumentation/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  Http: require('./http')
+}
